refactor:カスタムフックにロジックを移動
